+++ tests/testthat/test-wb_functions.R
@@ -262,3 +262,23 @@ test_that("improve date detection", {
   expect_equal(exp, got)
 
 })
+
+test_that("skip hidden columns and rows works", {
+
+  wb <- wb_workbook()$
+    add_worksheet()$
+    add_data(x = mtcars)$
+    set_col_widths(cols = c(1, 4, 6, 7, 9), hidden = TRUE)$
+    set_row_heights(rows = c(3, 5, 8:30), hidden = TRUE)
+
+  dat <- wb_to_df(wb, skipHiddenRows = TRUE, skipHiddenCols = TRUE)
+
+  exp <- c("2", "4", "6", "7", "31", "32", "33")
+  got <- rownames(dat)
+  expect_equal(exp, got)
+
+  exp <- c("cyl", "disp", "drat", "vs", "gear", "carb")
+  got <- names(dat)
+  expect_equal(exp, got)
+
+})